+++ offchain/scheduler/CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-- feat(automate_calculate_distribution): Add GenServer and Rust NIF to automatically calculuae a distribution on a configurable interval ([#197]https://github.com/doublezerofoundation/doublezero-offchain/pull/197)
+- feat(automate_finalize_distribution):
+- feat(automate_calculate_distribution): Add GenServer and Rust NIF to automatically calculate a distribution on a configurable interval ([#197]https://github.com/doublezerofoundation/doublezero-offchain/pull/197)
 - feat(automate_initialize_distribution): Add GenServer and Rust NIF to automatically initialize a distribution on a configurable interval ([#197]https://github.com/doublezerofoundation/doublezero-offchain/pull/197)
 - feat(automate_debt_payment): Add GenServer and Rust NIF to automatically collect debt on a configurable interval ([#183]https://github.com/doublezerofoundation/doublezero-offchain/pull/183)
 - feat(scheduler): add Elixir app that manages scheduling and executing Rust processes for debt collection and payment  ([#183]https://github.com/doublezerofoundation/doublezero-offchain/pull/183)